# public/swagger.yaml
openapi: 3.0.3
info:
  title: Air5Star eCommerce AI-Powered API
  description: |
    AI-powered features API for Air5Star HVAC eCommerce platform.
    
    This API provides intelligent features including:
    - Smart chatbot integration with product lookup
    - Advanced search with autocomplete and highlighting
    - Personalized product recommendations
    - Abandoned cart email automation
    - Inventory forecasting and alerts
    
    ## Security
    - All endpoints use HTTPS
    - API keys stored in environment variables
    - HMAC signature verification for webhooks
    - Minimal PII exposure to third-party services
    
    ## Rate Limiting
    - Search endpoints: 100 requests/minute
    - Recommendation endpoints: 50 requests/minute
    - Webhook endpoints: 1000 requests/hour
  version: 1.0.0
  contact:
    name: Air5Star Development Team
    email: dev@air5star.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://air5star.com/api
    description: Production server
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticate user with email/phone and password.
        Returns user data and JWT token for authenticated sessions.
        Supports both email and mobile number login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emailOrMobile
                - password
              properties:
                emailOrMobile:
                  type: string
                  description: User's email address or mobile number
                  example: "admin@hvac.com"
                password:
                  type: string
                  format: password
                  description: User's password
                  example: "admin123"
            examples:
              email_login:
                summary: Login with email
                value:
                  emailOrMobile: "admin@hvac.com"
                  password: "admin123"
              phone_login:
                summary: Login with phone
                value:
                  emailOrMobile: "+1234567890"
                  password: "userpassword"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT authentication token
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              example:
                message: "Login successful"
                user:
                  id: "1"
                  email: "admin@hvac.com"
                  name: "Admin User"
                  role: "ADMIN"
                  phone: "+1234567890"
                  createdAt: "2024-01-01T00:00:00Z"
                  updatedAt: "2024-01-01T00:00:00Z"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Please enter a valid email address or mobile number"
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                examples:
                  invalid_credentials:
                    summary: Invalid credentials
                    value:
                      error: "Invalid credentials"
                  oauth_user:
                    summary: OAuth user trying password login
                    value:
                      error: "Please login using your social account"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to login"
components:
  securitySchemes:
    HmacSignature:
      type: apiKey
      in: header
      name: X-Tidio-Signature
      description: HMAC SHA-256 signature for webhook verification
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's full name
        role:
          type: string
          enum: ["USER", "ADMIN"]
          description: User's role in the system
        phone:
          type: string
          description: User's phone number
        provider:
          type: string
          description: Authentication provider (local, google, etc.)
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        id: "1"
        email: "admin@hvac.com"
        name: "Admin User"
        role: "ADMIN"
        phone: "+1234567890"
        provider: "local"
        createdAt: "2024-01-01T00:00:00Z"
        updatedAt: "2024-01-01T00:00:00Z"
...